# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CandidateInput {
  name: string;
  skills: string;
  experience: string;
  education: string;
  location: string;
  currentCompany: string;
}

export type ScoredCandidateStatus =
  (typeof ScoredCandidateStatus)[keyof typeof ScoredCandidateStatus];

export const ScoredCandidateStatus = {
  QUALIFIED: "QUALIFIED",
  REVIEW: "REVIEW",
  REJECTED: "REJECTED",
} as const;

export interface ScoredCandidate {
  name: string;
  skills: string;
  experience: string;
  education: string;
  location: string;
  currentCompany: string;
  score: number;
  status: ScoredCandidateStatus;
  skillsScore: number;
  experienceScore: number;
  educationScore: number;
  locationScore: number;
  message?: string | null;
}

export interface ScoreCandidatesRequest {
  candidates: CandidateInput[];
}

export interface PipelineStats {
  total: number;
  qualified: number;
  review: number;
  rejected: number;
  timeSavedMinutes: number;
}

export interface ScoreCandidatesResponse {
  candidates: ScoredCandidate[];
  stats: PipelineStats;
}

export interface GenerateMessageRequest {
  name: string;
  skills: string;
  experience: string;
  currentCompany: string;
}

export interface GenerateMessageResponse {
  message: string;
}

export interface SendToSheetsRequest {
  name: string;
  score: number;
  skills: string;
  status: string;
  message?: string | null;
}

export interface SendToSheetsResponse {
  success: boolean;
  message: string;
}