# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Score candidates from CSV data
 */
export const ScoreCandidatesBody = zod.object({
  candidates: zod.array(
    zod.object({
      name: zod.string(),
      skills: zod.string(),
      experience: zod.string(),
      education: zod.string(),
      location: zod.string(),
      currentCompany: zod.string(),
    }),
  ),
});

export const ScoreCandidatesResponse = zod.object({
  candidates: zod.array(
    zod.object({
      name: zod.string(),
      skills: zod.string(),
      experience: zod.string(),
      education: zod.string(),
      location: zod.string(),
      currentCompany: zod.string(),
      score: zod.number(),
      status: zod.enum(["QUALIFIED", "REVIEW", "REJECTED"]),
      skillsScore: zod.number(),
      experienceScore: zod.number(),
      educationScore: zod.number(),
      locationScore: zod.number(),
      message: zod.string().nullish(),
    }),
  ),
  stats: zod.object({
    total: zod.number(),
    qualified: zod.number(),
    review: zod.number(),
    rejected: zod.number(),
    timeSavedMinutes: zod.number(),
  }),
});

/**
 * @summary Generate personalized LinkedIn outreach message using Gemini
 */
export const GenerateMessageBody = zod.object({
  name: zod.string(),
  skills: zod.string(),
  experience: zod.string(),
  currentCompany: zod.string(),
});

export const GenerateMessageResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Send qualified candidate data to Google Sheets webhook
 */
export const SendToSheetsBody = zod.object({
  name: zod.string(),
  score: zod.number(),
  skills: zod.string(),
  status: zod.string(),
  message: zod.string().nullish(),
});

export const SendToSheetsResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});