# lib/api-zod/src/generated/types/generateMessageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateMessageResponse {
  message: string;
}