# lib/api-zod/src/generated/types/pipelineStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PipelineStats {
  total: number;
  qualified: number;
  review: number;
  rejected: number;
  timeSavedMinutes: number;
}