# lib/api-zod/src/generated/types/scoreCandidatesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateInput } from "./candidateInput";

export interface ScoreCandidatesRequest {
  candidates: CandidateInput[];
}