# lib/api-zod/src/generated/types/scoreCandidatesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PipelineStats } from "./pipelineStats";
import type { ScoredCandidate } from "./scoredCandidate";

export interface ScoreCandidatesResponse {
  candidates: ScoredCandidate[];
  stats: PipelineStats;
}