# lib/api-zod/src/generated/types/scoredCandidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecruitAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScoredCandidateStatus } from "./scoredCandidateStatus";

export interface ScoredCandidate {
  name: string;
  skills: string;
  experience: string;
  education: string;
  location: string;
  currentCompany: string;
  score: number;
  status: ScoredCandidateStatus;
  skillsScore: number;
  experienceScore: number;
  educationScore: number;
  locationScore: number;
  message?: string | null;
}